refactor: move the initialization code to `endpoint_initializer.rs`
